fix(form): enforce integer input validation for stats
Update validation rules from 'numeric' to 'integer' in request class
Add client-side validation to prevent non-integer input
Enhance input handling with paste event filtering and min/max enforcement

+++ app/Requests/StatCalculatorRequest.php
@@ -18,13 +18,13 @@ class StatCalculatorRequest extends FormRequest
         return [
             'agility' => [
                 'required',
-                'numeric',
+                'integer',  // Changed from 'numeric' to 'integer'
                 'min:0',
                 'max:9999',
             ],
             'strength' => [
                 'required',
-                'numeric', 
+                'integer', 
                 'min:0',
                 'max:9999',
             ],
